apps/worker/pipeline/requirements.txt
claude-agent-sdk>=0.1.48
anyio>=4.0
python-dotenv>=1.0
httpx>=0.27
mcp>=1.0.0,<2.0.0
